+++ extensions/amp-analytics/0.1/instrumentation.js
@@ -17,8 +17,7 @@
 import {AmpdocAnalyticsRoot, EmbedAnalyticsRoot} from './analytics-root';
 import {AnalyticsEvent, CustomEventTracker} from './events';
 import {AnalyticsGroup} from './analytics-group';
-import {Services} from '../../../src/services';
-import {getFriendlyIframeEmbedOptional} from '../../../src/friendly-iframe-embed';
+import {Services, getFriendlyIframeEmbedOptional} from '../../../src/services';
 import {
   getParentWindowFrameElement,
   getServiceForDoc,
